requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.4
pydantic==2.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic_core==2.16.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
